backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
mlflow>=2.8.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
sqlalchemy>=2.0.0
opencv-python>=4.8.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
mlflow>=2.8.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
sqlalchemy>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pydantic>=2.0.0
aiofiles>=23.0.0
requests>=2.31.0
pymongo>=4.6.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
email-validator>=2.1.0
